+++ nestjs-auth/src/auth-typeorm/auth.controller.ts
@@ -10,7 +10,7 @@ import {
   UseGuards,
 } from '@nestjs/common';
 import { AuthService } from './auth.service';
-import { LocalAuthGuard } from './guards/local-auth.guard';
+import { LocalAuthGuard } from './guard/local-auth.guard';
 
 @Controller('')
 export class AuthController {
